Fixed #5
Modified code to pass 'this' to the ShowDialg() calls for Export ALl and
Import.

Note on Linux systems, there's a small chance this won't completely fix
the problem, since window focus is handled at the window manager level.
However, this ensures that hints are set properly, so the window manager
has a chance to do the right thing.

Note that this problem was only intermittent for me, but that's to be
expected, as the window manager might set focus back to the main program
window, when lacking anything to tell it where to set focus.