Fixed #11
Modified the code to check if we're running on Windows or not.  If
running on Windows, use the existing code to open up Explorer and browse
to the folder.  Otherwise, attempt to execute the folder itself.  In
most Mono implementations, this will actually do the right thing, and
bring up the default file browser showing the folder.