Added new icon for KFileDialog and hooked up refresh button
Added a new icon for KFileDialog for its view button.  Looks vaguely
like MS's view button icon, so it should be recognizable.

Also added functionality for the refresh button.  When you click it,
it'll refresh the list of files in the dialog.  It doesn't refresh the
list of mounted directories, not sure if it should or not.

Fixed #42.