Fixed #57
Added code to set this.DialogResult to DialogResult.OK when the OK
button is clicked.  This seems to resolve the problem.

I'm not sure how this was working before, or how it suddenly stopped
working.  This change should guarantee it works as expected (pending any
changes from upstream that may test the result of the dialog
differently).