Fixed #55
Changed code to use a regular ListView for Mono instead of
FastObjectListView.  The FastObjectListView doesn't seem to work on
Mono, AFAICT, it never tries to display the information at all, and my
best-guess is that's due to a bug in Mono itself.  None of the special
features of ObjectListView or FastObjectListView are used, so the only
possible advantage is if they are less limited in the number of errors
they can display, than is ListView.  There's a small chance that
ListView will fail (i.e. not display them all) if there are too many
errors, but at that point, I'm not sure the error information will be
useful, since it'll be many thousands of errors (the old limit on
Windows was 32,767, but that was doubled, IIRC in Win2k, XP and later,
and I've no idea what Mono's limit might be).  If the limitation is
causing issues, I'll probably rewrite the dialog completely. :)